Ros2_CNC
I designed a cnc robot and added a camera and lidar to this robot. The camera can select red blocks. Lidar detects the distance of the objects I put. I can transfer the distance I get with the lidar to the robot and go to that point with the robot.